+++ src/chattice/__init__.py
@@ -5,4 +5,4 @@ from .filters import F
 
 __all__ = ["Dispatcher", "F", "Router", "__version__"]
 
-__version__ = "0.14.0b5"
+__version__ = "0.14.0b6"